Backend/requirements.txt
Flask
pymongo
Flask-PyMongo
Werkzeug
python-dotenv
Flask-CORS
PyJWT
bcrypt
Flask-SQLAlchemy
requests
supabase
